fix sync errors
