+++ packages/query/src/block-processor.ts
@@ -1,7 +1,7 @@
 import { RootQuerier } from './root-querier';
 
 import { sha256Hash } from '@penumbra-zone/crypto-web';
-import { computePositionId, decodeSctRoot } from '@penumbra-zone/wasm';
+import { computePositionId, decodeSctRoot, getLpNftMetadata } from '@penumbra-zone/wasm';
 import {
   PositionState,
   PositionState_PositionStateEnum,
@@ -297,17 +297,16 @@ export class BlockProcessor implements BlockProcessorInterface {
 
   private async identifyLpNftPositions(txs: Transaction[]) {
     const positionStates: PositionState[] = [
-      new PositionState({state: PositionState_PositionStateEnum.OPENED}),
-      new PositionState({state: PositionState_PositionStateEnum.CLOSED}),
-      new PositionState({state: PositionState_PositionStateEnum.CLOSED, sequence: 0n}),
+      new PositionState({ state: PositionState_PositionStateEnum.OPENED }),
+      new PositionState({ state: PositionState_PositionStateEnum.CLOSED }),
+      new PositionState({ state: PositionState_PositionStateEnum.CLOSED, sequence: 0n }),
     ];
 
     for (const tx of txs) {
-      console.log(tx)
       for (const { action } of tx.body?.actions ?? []) {
         if (action.case === 'positionOpen' && action.value.position) {
           for (const state of positionStates) {
-            const metadata = this.viewServer.getLpNftMetadata(action.value.position, state);
+            const metadata = getLpNftMetadata(computePositionId(action.value.position), state);
             await this.indexedDb.saveAssetsMetadata(metadata);
           }
           // to optimize on-chain storage PositionId is not written in the positionOpen action,
@@ -325,14 +324,14 @@ export class BlockProcessor implements BlockProcessorInterface {
         }
         if (action.case === 'positionWithdraw' && action.value.positionId) {
           // Record the LPNFT for the current sequence number.
-          let positionState = new PositionState({ state: PositionState_PositionStateEnum.WITHDRAWN, sequence: action.value.sequence});
-          const metadata = this.viewServer.getLpNftMetadata(action.value.position, positionState);
+          let positionState = new PositionState({
+            state: PositionState_PositionStateEnum.WITHDRAWN,
+            sequence: action.value.sequence,
+          });
+          const metadata = getLpNftMetadata(action.value.positionId, positionState);
+          await this.indexedDb.saveAssetsMetadata(metadata);
 
-
-          await this.indexedDb.updatePosition(
-            action.value.positionId,
-            positionState,
-          );
+          await this.indexedDb.updatePosition(action.value.positionId, positionState);
         }
       }
     }
